feat: nlp embeding logick

+++ init.sql
@@ -8,6 +8,6 @@ CREATE TABLE IF NOT EXISTS documents(
     id SERIAL PRIMARY KEY,
     title TEXT NOT NULL,
     content TEXT NOT NULL,
-    embedding INTEGER[],
+    embedding FLOAT[],
     created_at TIMESTAMP
 );
